+++ packages/gasket-start-plugin/CHANGELOG.md
@@ -1,5 +1,10 @@
 # `@gasket/start-plugin`
 
+- Allow env to be set for **local** command, with a default of `local` ([#91])
+
 ### 1.0.0
 
 - Initial implementation
+
+
+[#91]: https://github.com/godaddy/gasket/pull/91